refactor(aggregator): move `GenesisTools` back to its original location, keeping only `GenesisToolsDependency` in the dependency module
